Add ORDER_BY support for AST collection nodes (#30)
Adds a collection node for ORDER_BY clauses which is derived from the same base class as `Calc` and `Where`. Instead of .with_terms, it has a .with_collation method that must be called. The collation is a list of `CollationExpression` nodes, a new type of AST expression node that wraps around another expression and has asc/desc + nullpos info.