Refactoring collections to separate ChildAccess from CollectionAccess (#34)
Adding a general `ChildAccess` class for any operator that steps down into a child, which `CollectionAccess` inherits from & is a special case where the accessing is via collection/subcollection metadata. This is necessary for when `PARTITION` is introduced. Also does some refactors to the stringification process to separate an individual subtree from its ancestor tree, allowing for better re-use of code & less duplication.