Add TOP K support for collection nodes (#35)
Adding support for a `TopK` AST node which is a subclass of an `OrderBy` node. This always has a collation specified via `.with_collation` just like an order by node, but in practice if one is not provided by the user we will derive it from the predecessor before creating the `TopK` node.