Reworking how AST nodes do hashing (#50)
Each AST node now has a `key` property that returns a string, which is not necessarily unique even for non-equal nodes (since it does not evaluate the lazy terms), this increasing the importance of equality checks. The hashing for an AST node extracts this string and hashes it via sha256 (possibly will change later).