Setting up the unqualified transform step (#45)
Adds an `unqualified_transform` step that uses a function `transform_code` to take in Python source code (as a string), the JSON metadata file path + graph name, and a set of known variable names (e.g. global names), and returns a Python AST corresponding to the transformed code. Also adds a decorator `init_pydough_context` that invokes this procedure on its function body and compiles the result. The transformations remove the decorator from the function, add the definition of a variable `ROOT` to the top of the function body, and prepends any undefined variables with `_ROOT.`

Augments all of the regular tests for the unqualified API by ensuring they return the same answer even when `_ROOT.` is removed from the strings. Also tests on TPCH queries 6 and 10 (located in another file where `mypy` is not run).