Begin AST->Relational conversion (#78)
First step in converting the PyDough AST nodes to Relational nodes. Creates a HybridTree structure where each level contains a pipeline of operations, optionally has a predecessor/successor level, and can contain a list of connections to children that are evaluated so that operations in the pipeline can make use of them. Also adds a function `convert_ast_to_relational` that converts a PyDough collection AST to this HybridTree form (first pre-processing it so the final node is a calc with the terms that will be needed for the final columns + ordering), then converting the HybridTree form to a Relational node tree by starting at the last element of the pipeline at the bottom of the tree. This function does so by recursively converting the preceding operation (either the previous element in the current pipeline or the last element in the previous pipeline, if at the start of the current pipeline) into a Relational node along with extra information about how to access expressions form the hybrid tree from this node, then applying operation-specific logic to augment all of that information & obtain the new relational node + expression access logic.

Currently only supports basic table scans, basic `collection.subcollection` accesses, and basic calc terms with literals, columns, and references (no functions). Does not yet support any child accesses (e.g. `Nations(region_name=region.name)`, or any aggregation).