Support basic singular child accesses as LEFT joins in AST->Relational conversion (#95)
Initial support for converting AST nodes to Relational nodes when they contain singular child references, such as `Customers(region_name=nation.region.name)`. Done by creating an alternative path for conversion to process the child nodes of a HybridTree without stepping into the context, and then doing a left-join instead of an inner-join to reconcile the two halves. For now, the two sides must be independent, so having the child nodes depend on the current context or its ancestors in any way is not possible.