Support basic aggregations in AST->Relational conversion (#98)
Supports simple aggregations in AST to Relational conversion, such as `Nations(tot_sup_value=SUM(suppliers.account_balance))`. Done by augmenting the `HybridConnection` instance to store a dictionary of aggregation calls (when the connection type is `AGGREGATION`). Then, during the `HybridTree` conversion step, aggregation calls are converted to instead be a regular child reference, and when the relational translation step is done, HybridChildren that are `AGGREGATION` are post-processed before the left-join by adding an `AGGREGATE` node using the RHS join keys as the grouping keys and using the aggregates stored in the `HybridConnection`. Note: this only works if the arguments to the aggregation depend on a **single** child reference.

Also adds the initial support for PyDough configurations, with 2 configs added for the behaviors of SUM and AVG when there are no records to aggregate. In this case, the default behavior for both configs is to coalesce the output with 0.