Enable runtime correctness testing of TPCH queries (#115)
Changed the tests in `test_pipeline.py` to also include the tpch output refsols, and to execute the query after converting to relational form + convert to the answer. Currently has running answers for all except the 3 queries that are not supported in AST->Relational conversion. Also added a pytest marker `execute` for tests that do runtime execution so they can be skipped during local development with `-m "not slow"`.