Added graphs for defog.ai schemas (#192)
Graphs were generated using Google Gemini, fed with the spec for PyDough JSON metadata and previously written semantic information about the Ewallet, DermTreatment, Dealership, and Broker schemas. Also adding a script to populate a local sqlite database with the corresponding tables (so far only for Broker) & adding e2e tests to run defog.ai queries with PyDough and compare the outputs to the SQLite query implementation (so far only for 3 of the broker queries).