Update column name generation in relational conversion (#244)
Ancillary task related to #141. Changes how column names are handled in CALC so existing names are not overriden, but instead a new name is used. This becomes crucial for when correlated references are implemented, but until then causes harmless naming changes to many relational plans. These changes are benign because they are not observable in the final answer, and only change the internal names.