Support uses of BACK that cause correlated references: setup decorrelation handling (#251)
Child PR of #269 that is part of addressing #141. Adds the handling of de-correlation as a post-processing step of the hybrid tree before relational conversion starts. Handles the singular & aggregate cases as described in the issue, and handles the semi-singular and semi-aggregate cases in the same manner as their non-semi equivalents (for now), saving the optimized variants for another day.

These remaining still need to be addressed in the next follow up:
- TPC-H Q22: error during relational conversion that likely means a bug happened during hybrid conversion or hybrid decorrelation
- Correl 15: error during decorrelation causing over-excessive column pruning
- Correl 18: tbd bug in decorrelation
- Correl 19: bug in name resolution handling that causes the wrong term to be used when certain property names collide