Fixing hybrid conversion bug that caused naming errors (#283)
The bug was brought to attention by cases like this:

```py
urgent_orders = (
  customers.CALCULATE(nation_name=nation.name)
  .orders
  .WHERE(order_priority == "1-URGENT")
  .CALCULATE(nation_name, order_year=YEAR(order_date), order_month=MONTH(order_date))
)
groups = PARTITION(urgent_orders, name="u", by=(nation_name, order_year, order_month))
result = groups.CALCULATE(nation_name, order_year, order_month, n_orders=COUNT(u))
```

There is a name-resolution bug during hybrid conversion that causes `nation_name` to be incorrectly changed to `nation_name_0` when used as a child reference to the data being partitioned, which causes a key error during relational conversion when it looks for `$0.nation_name_0` in the expression dictionary but can't find it.

The fix is to avoid the invocation of `nation_name` in the second `CALCULATE` causing an extra renaming since it is just a re-invocation of the same term.