Optimize PyDough generated SQL to avoid extra joins from PARTITION when possible (#289)
Consider this PyDough logic:
```py
priorities = PARTITION(
   Orders.WHERE(YEAR(order_date) == 1992), name="orders", by=order_priority
)
result = (
   priorities.orders.WHERE(RANKING(by=total_price.DESC(), levels=1) == 1)
   .CALCULATE(order_priority, order_key=key, order_total_price=total_price)
   .ORDER_BY(order_priority.ASC())
)
```

The goal is to ensure that the final SQL does **NOT** derive `Orders.WHERE(YEAR(order_date) == 1992)` twice and join them. Here is the relational plan before implementing the change:

```
ROOT(columns=[('order_priority', order_priority_4), ('order_key', order_key), ('order_total_price', order_total_price)], orderings=[(ordering_0):asc_first])
 PROJECT(columns={'order_key': order_key, 'order_priority_4': order_priority_3, 'order_total_price': order_total_price, 'ordering_0': ordering_0})
  PROJECT(columns={'order_key': order_key, 'order_priority_3': order_priority_3, 'order_total_price': order_total_price, 'ordering_0': order_priority_3})
   PROJECT(columns={'order_key': key, 'order_priority_3': order_priority_1, 'order_total_price': total_price})
    FILTER(condition=RANKING(args=[], partition=[order_priority], order=[(total_price):desc_first]) == 1:int64, columns={'key': key, 'order_priority_1': order_priority_1, 'total_price': total_price})
     JOIN(conditions=[t0.order_priority == t1.order_priority], types=['inner'], columns={'key': t1.key, 'order_priority': t0.order_priority, 'order_priority_1': t1.order_priority, 'total_price': t1.total_price})
      AGGREGATE(keys={'order_priority': order_priority}, aggregations={})
       FILTER(condition=YEAR(order_date) == 1992:int64, columns={'order_priority': order_priority})
        SCAN(table=tpch.ORDERS, columns={'order_date': o_orderdate, 'order_priority': o_orderpriority})
      FILTER(condition=YEAR(order_date) == 1992:int64, columns={'key': key, 'order_priority': order_priority, 'total_price': total_price})
       SCAN(table=tpch.ORDERS, columns={'key': o_orderkey, 'order_date': o_orderdate, 'order_priority': o_orderpriority, 'total_price': o_totalprice})
```

Notice how the subtree that filters orders is derived twice: once aggregated on order priority, then joined with the original. This is needed in general with the `PARTITION(...).x` pattern since the partition could have aggregated its original data (so we can compare rows of `x` to the aggregates) or filtered the partitioned data (e.g. finding singleton partitions, or top groups, etc.) so the `.x` access is effectively filtering to only include rows of `x` that meet certain window-esque criteria. However, it is **NOT** needed in this case because the data of the two subtrees is identical (except for the aggregation, which is also used as the join key).

The final plan **should** look like this:

```
ROOT(columns=[('order_priority', order_priority_4), ('order_key', order_key), ('order_total_price', order_total_price)], orderings=[(ordering_0):asc_first])
 PROJECT(columns={'order_key': order_key, 'order_priority_4': order_priority_3, 'order_total_price': order_total_price, 'ordering_0': ordering_0})
  PROJECT(columns={'order_key': order_key, 'order_priority_3': order_priority_3, 'order_total_price': order_total_price, 'ordering_0': order_priority_3})
   PROJECT(columns={'order_key': key, 'order_priority_3': order_priority, 'order_total_price': total_price})
    FILTER(condition=RANKING(args=[], partition=[order_priority], order=[(total_price):desc_first]) == 1:int64, columns={'key': key, 'order_priority': order_priority, 'total_price': total_price})
     FILTER(condition=YEAR(order_date) == 1992:int64, columns={'key': key, 'order_priority': order_priority, 'total_price': total_price})
      SCAN(table=tpch.ORDERS, columns={'key': o_orderkey, 'order_date': o_orderdate, 'order_priority': o_orderpriority, 'total_price': o_totalprice})
```

The implementation of this optimization is done via a special edge-case in the relational conversion of a PartitionChild access. If the specified pattern is detected before joining (meaning context is just an aggregate with no aggregations), then the child output is the answer (with its expressions slightly shifted) and the context can be ignored.