Automatically update defog.db whenever tests are run (#329)
Ensures that the `setup_defog.sh` script is run whenever the defog database is going to be used by the unit tests within the context of the pytest command. This ensures the relative datetimes are always up to date, and any new tables in the `init_defog.sql` script are automatically added on other branches without needing to re-run the script manually.