Fixing bug with test broker_gen2 (#392)
The actual defog SQL text is as follows:
```sql
SELECT COUNT(t.sbTxCustId) AS transaction_count
FROM sbTransaction AS t
JOIN sbCustomer AS c ON t.sbTxCustId = c.sbCustId
WHERE c.sbCustJoinDate >= date('now', '-70 days')
```
But the current PyDough implemnetation uses `DATETIME("now", "-70 days")` which does not truncate to the start of the day. To fix this, changed it to `DATETIME("now", "-70 days", "start of day")` so it is the same as the refsol.