Improve inference of predicates on top of joins (#397)
Extension of #396 to further the transformation of LEFT joins into INNER joins when there are filters on top of the join if the filters would be false if the RHS columns were all-null (relies on simplification to deduce if this is the case). Also creates a new `RelationalShuttle` base class so relational nodes can use the shuttle design pattern, and refactors filter pushdown to use this pattern. Also refactors several other utilities to use shuttles that should have been doing so already.