Add masked table column relational unprotect rewrite (#417)
Adding crucial rewrite step for masked table columns: During relational conversion, if a column is a masked table column, place a `PROJECT` on top of the `SCAN` node fetching data from the table. This `PROJECT` node will invoke an `UNMASK` operator (containing information from the metadata for the masked table column) which will transform the masked columns into their unmasked forms.

Updates existing tests to account for this transformation, ensuring that the `UNMASK` calls are injected into the underlying SQL (and pulled up as late as possible by projection pullup). All the tests in `test_masked_sqlite.py` will now run with the _correct_ e2e answers, as the decryption of the underlying data is now undone by the generated query.