Delete redundant joins by substitution equi-join keys for their mirror to render one side pruneable (#450)
Adds an optimization pass which substitutes columns used by a join from one side to the other when passing through a join key. This is done under a certain set of conditions to ensure that the one side of the join no longer has any of its columns being used, thus ensuring that under those same conditions, the join can be optimized out by the column pruner.